chore: extract-method control_if
